backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
openai>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
